Moved symlinkPrivilegeWarnings() from utils to test-utils. (#2590)
