+++ packages/core/utils/src/errorUtils.js
@@ -27,14 +27,5 @@ function jsonToError(json) {
   }
 }
 
-function symlinkPrivilegeWarning() {
-  console.error("-----------------------------------")
-  console.error("Skipping symbolic link test(s) because you don't have the privilege.");
-  console.error("Run tests with Administrator privilege.");
-  console.error("If you don't know how, check here: https://bit.ly/2UmWsbD");
-  console.error("-----------------------------------")
-}
-
 exports.errorToJson = errorToJson;
 exports.jsonToError = jsonToError;
-exports.symlinkPrivilegeWarning = symlinkPrivilegeWarning;
